requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
nltk==3.8.1
python-dotenv==1.0.0
pytest==7.4.2